# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_product_api_v1_store_products_upload_post": {
        "properties": {
          "asset": {
            "contentMediaType": "application/octet-stream",
            "title": "Asset",
            "type": "string"
          },
          "metadata": {
            "title": "Metadata",
            "type": "string"
          }
        },
        "required": [
          "metadata",
          "asset"
        ],
        "title": "Body_upload_product_api_v1_store_products_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StoreAssetResponse": {
        "properties": {
          "assetType": {
            "title": "Assettype",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mimeType": {
            "title": "Mimetype",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "previewKind": {
            "title": "Previewkind",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "assetType",
          "mimeType",
          "sizeBytes",
          "sha256",
          "previewKind"
        ],
        "title": "StoreAssetResponse",
        "type": "object"
      },
...
      "StoreProductResponse": {
        "properties": {
          "acgIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Acgids",
            "type": "array"
          },
          "areaOrRegion": {
            "title": "Areaorregion",
            "type": "string"
          },
          "assets": {
            "items": {
              "$ref": "#/components/schemas/StoreAssetResponse"
            },
            "title": "Assets",
            "type": "array"
          },
          "classificationLevel": {
            "title": "Classificationlevel",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "geojsonRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geojsonref"
          },
          "handlingCaveats": {
            "items": {
              "type": "string"
            },
            "title": "Handlingcaveats",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ownerTeam": {
            "title": "Ownerteam",
            "type": "string"
          },
          "productType": {
            "title": "Producttype",
            "type": "string"
          },
          "reference": {
            "title": "Reference",
            "type": "string"
          },
          "releasability": {
            "items": {
              "type": "string"
            },
            "title": "Releasability",
            "type": "array"
          },
          "semanticLabels": {
            "items": {
              "type": "string"
            },
            "title": "Semanticlabels",
            "type": "array"
          },
          "sourceType": {
            "title": "Sourcetype",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "timePeriodEnd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeperiodend"
          },
          "timePeriodStart": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeperiodstart"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reference",
          "title",
          "summary",
          "description",
          "productType",
          "sourceType",
          "ownerTeam",
          "areaOrRegion",
          "classificationLevel",
          "releasability",
          "handlingCaveats",
          "tags",
          "semanticLabels",
          "acgIds",
          "status",
          "timePeriodStart",
          "timePeriodEnd",
          "geojsonRef",
          "assets"
        ],
        "title": "StoreProductResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Secure intelligence tasking and product orchestration API.",
    "title": "Istari API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/store/products/upload": {
      "post": {
        "operationId": "upload_product_api_v1_store_products_upload_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_product_api_v1_store_products_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoreProductResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Product",
        "tags": [
          "store"
        ]
      }
    },
...
    "/api/v1/store/products/{product_id}/assets/{asset_id}/download": {
      "get": {
        "operationId": "download_asset_api_v1_store_products__product_id__assets__asset_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "product_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Product Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "asset_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Asset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Asset-Token",
            "required": true,
            "schema": {
              "minLength": 20,
              "title": "X-Asset-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Asset",
        "tags": [
          "store"
        ]
      }
    },
...
  }
}